requirements.txt
aws-cdk-lib==2.10.0
boto3==1.15.16
botocore==1.18.16

# packages for aws lambda layer
# DO NOT REMOVE [[layers]] comment
#[[layers]]
arrow==0.17.0
requests==2.24.0
beautifulsoup4==4.9.3
newspaper3k==0.2.8
redis==3.3.11

